sched/bore: cap demotion at +4 nice, fix u32 sum overflow
Two fixes to stop BORE's unbounded ratcheting over long sessions:

MAX_BURST_PENALTY 156 -> 16 (burst_score cap 39 -> 4). The ratchet only
releases at sleep/yield, so a thread that stays runnable across a long
game session accumulates weight loss with no floor: the old cap scheduled
the main thread at up to nice +39 by the back of the session. A +4 cap
keeps batch work clearly demoted (enough for EAS packing) but bounds a
long session to nice +4. u8 KABI fields unaffected: 16 < 255.

update_child_burst_topological(): the u8 x u32 product in the cached
subtree sum widened to u32 and could wrap at large child_burst_cnt.
Widen to u64 before the cast back to u32.

The topological walk is dead code (fork_atavistic pinned to 0 via
sysctl clamps), so the u64 change is type-correctness only and does not
alter runtime behaviour. The cap is the functional fix.

No change to the measured values: offset 27, scale 1024, smoothness
1/0, fork_atavistic 0, cache lifetime 75ms all left untouched.

(cherry picked from commit cacc6ef50a77b4b5251d32d55e4d259ec9f49a1b)

+++ kernel/sched/core.c
@@ -3369,7 +3369,9 @@ static void update_child_burst_topological(
 		}
 		if (!child_burst_cache_expired(dec, now)) {
 			cnt += dec->se.child_burst_cnt;
-			sum += (u32)dec->se.child_burst * dec->se.child_burst_cnt;
+			/* u8 x u32 widens to u32 and wraps; keep the product in u64. */
+			sum += (u32)((u64)dec->se.child_burst *
+					 dec->se.child_burst_cnt);
 			continue;
 		}
 		update_child_burst_topological(dec, now, depth - 1, &cnt, &sum);
